requirements.txt
aiogram>=3.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
 
# Указать минимальную версию Python
python_version >= "3.10" 